test: Rename 'tests' to 'test'
Move test harness include files to a more logical location. Also renames
'harness.h' to 'test.h' to better reflect what's in it.
