+++ src/test/random.cu
@@ -29,7 +29,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <lift/tests/random.h>
+#include <lift/test/random.h>
 
 #include <random>
 #include <cmath>
